change logger.ts
